move: 패키지 이동
- common/util -> common/presentation/util
